Some tweaks.
git-svn-id: svn+ssh://clio.psy.vu.nl/home/r.pool/svn/BioCrates/trunk@105 f67a8107-f8d1-4425-a0ae-94eea21ecd1d

+++ MetaAnalysis/BoxPlot/BoxPlot.py
@@ -223,7 +223,7 @@ def main(ExecutableName=str):
 #       TwinsUK scaling by 1/log10(e)
         TwinsUKScalingFactor = None
         if(eval(XmlProtocol.getroot().find('Scaling').find('UseTwinsUKScaling').text)):
-            TwinsUKScalingFactor = 1.0/scipy.log10(scipy.exp(1.0))
+            TwinsUKScalingFactor = float(1.0/scipy.log10(scipy.exp(1.0)))
 
         LogString = '  ++ Plotting \"SE\" box plot ...'
         print LogString
